requirements.txt
blinker==1.7.0
bootstrap-flask==2.5.0
click==8.1.7
flask-avatars==0.2.3
flask-dropzone==2.0.0
flask-login==0.6.3
flask-mail==0.9.1
flask-moment==1.0.6
flask-sqlalchemy==2.5.1
flask-whooshee==0.7.0
flask-wtf==1.2.1
flask==2.2.5
itsdangerous==2.2.0
jinja2==3.1.2
markupsafe==2.1.3
Pillow==11.0.0
python-dotenv==1.0.0
sqlalchemy==1.4.54
werkzeug==2.2.3
whoosh==2.7.4
wtforms==3.1.1
faker==4.0.2
pathtools==0.1.2    
python-dateutil==2.8.1
six==1.14.0
text-unidecode==1.3
watchdog==3.0.0
